refactor: rename uauth convention to auth
