+++ src/app/(auth)/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignupForm from "@/components/user-auth/signup-form";
+import SignupForm from "@/components/auth/signup-form";
 
 export default function Signup() {
   return (
